SDR Reshape class: moved from SdrTools.hpp to Sdr.hpp
Removed file SdrTools.
