+++ src/nupic/types/Sdr.cpp
@@ -540,5 +540,64 @@ namespace sdr {
         destroyCallbacks[index] = nullptr;
     }
 
+
+    /**************************************************************************/
+
+    Reshape::Reshape(const SDR &sdr, const vector<UInt> &dimensions)
+        : SDR( dimensions )
+    {
+        clear();
+        parent = &sdr;
+        NTA_CHECK( size == parent->size ) << "SDR Reshape must have same size as given SDR.";
+        callback_handle = parent->addCallback( [&] () {
+            clear();
+            do_callbacks();
+        });
+        destroyCallback_handle = parent->addDestroyCallback( [&] () {
+            deconstruct();
+        });
+    }
+
+    SDR_dense_t& Reshape::getDense() const {
+        NTA_CHECK( parent != nullptr ) << "Parent SDR has been destroyed!";
+        return parent->getDense();
+    }
+
+    SDR_sparse_t& Reshape::getSparse() const {
+        NTA_CHECK( parent != nullptr ) << "Parent SDR has been destroyed!";
+        return parent->getSparse();
+    }
+
+    SDR_coordinate_t& Reshape::getCoordinates() const {
+        NTA_CHECK( parent != nullptr ) << "Parent SDR has been destroyed!";
+        if( dimensions.size() == parent->dimensions.size() &&
+            equal( dimensions.begin(), dimensions.end(),
+                   parent->dimensions.begin() )) {
+            // All things equal, prefer reusing the parent's cached value.
+            return parent->getCoordinates();
+        }
+        else {
+            // Don't override getCoordinates().  It will call either getDense()
+            // or getSparse() to get its data, and will use this SDR Reshape's
+            // dimensions.
+            return SDR::getCoordinates();
+        }
+    }
+
+    void Reshape::save(std::ostream &outStream) const {
+        NTA_CHECK( parent != nullptr ) << "Parent SDR has been destroyed!";
+        parent->save( outStream );
+    }
+
+    void Reshape::deconstruct() {
+        // Unlink this SDR from the parent SDR.
+        if( parent != nullptr ) {
+            parent->removeCallback( callback_handle );
+            parent->removeDestroyCallback( destroyCallback_handle );
+            parent = nullptr;
+            SDR::deconstruct();
+        }
+    }
+
 } // end namespace sdr
 } // end namespace nupic
