+++ src/test/CMakeLists.txt
@@ -96,7 +96,6 @@ set(regions_tests
 set(types_tests
 	   unit/types/ExceptionTest.cpp
 	   unit/types/SdrTest.cpp
-	   unit/types/SdrToolsTest.cpp
 	   )
 	   
 set(utils_tests
